data types for writing drive velocity and reading velocity fixed

+++ Sources/rte/rte.c
@@ -335,13 +335,13 @@ StdRtn_t RTE_Play_BuzBeep(uint16 freqHz_, uint16 durMs_)
 /**
  * Interface implementation for the speedometer
  */
-StdRtn_t RTE_Read_SpdoVelLe(uint16 *vel_)
+StdRtn_t RTE_Read_SpdoVelLe(int16_t *vel_)
 {
 	return TACHO_Read_SpdLft(vel_);
 }
 
 
-StdRtn_t RTE_Read_SpdoVelRi(uint16 *vel_)
+StdRtn_t RTE_Read_SpdoVelRi(int16_t *vel_)
 {
 	return TACHO_Read_SpdRght(vel_);
 }
@@ -352,9 +352,9 @@ StdRtn_t RTE_Read_SpdoVelRi(uint16 *vel_)
 /**
  * Interface implementation for the drive component
  */
-StdRtn_t RTE_Write_DrvVel(int32 velLe_, int32 velRi_)
+StdRtn_t RTE_Write_DrvVel(int16_t velLe_, int16_t velRi_)
 {
-	return (StdRtn_t)DRV_SetSpeed(velLe_, velRi_);
+	return (StdRtn_t)DRV_SetSpeed((int32_t)velLe_, (int32_t)velRi_);
 }
 
 StdRtn_t RTE_Write_DrvPos(int32 posLe_, int32 posRi_)
